feat(allwinner): add A523 support
The Allwinner A523 features two clusters with four Cortex-A55 cores
each, also use an Arm GIC-600 for the interrupt controller. The memory
map changes a bit, but most of the peripherals are compatible as far as
TF-A is concerned.
An exception are the CPU ops, which turn the power of cores on or off,
they use a separate implementation, dictated by the DSU interconnect.

Add a new build platform, enabling the Cortex-A55 cores, and provide an
A55 specific Makefile snippet. Describe the new memory map, and add the
new CPU ops implementation.

Change-Id: I10d5bd1224e21593ab96ec62b986e01b141bf8d8
Signed-off-by: Jernej Skrabec <jernej.skrabec@gmail.com>
Signed-off-by: Andre Przywara <andre.przywara@arm.com>