feat(allwinner): plat_core_pos: support MT bit
So far all CPU cores used in Allwinner SoCs use the older MPIDR
affinity style, where the MT bit is clear, and the cores are enumerated
using the Aff0 MPIDR field.
Newer SoCs use a newer generation of Arm Cortex cores, which set the MT
bit and put the core number in the Aff1 field.

To accommodate both generations within the shared code, query the MT bit
during the calculation, and use one or the other calculation method.

Change-Id: I2aa5de8c759e0ae5d1cd4a04b6dcc82cadf855ef
Signed-off-by: Andre Przywara <andre.przywara@arm.com>